move specific wasi files into internal subfolder
