+++ std/assembly/bindings/internal/wasi_errors.ts
