+++ std/assembly/bindings/internal/wasi_snapshot_preview1.ts
