+++ std/assembly/bindings/internal/wasi_unstable.ts
