+++ std/assembly/bindings/wasi.ts
@@ -1,5 +1,5 @@
-export * from "./wasi_errors";
-export * from "./wasi_snapshot_preview1";
+export * from "./internal/wasi_errors";
+export * from "./internal/wasi_snapshot_preview1";
 
 // A WASI-wide reusable temporary buffer to store and work with out values. Must
 // be large enough to fit any operation it is used in, i.e. process/writeString.
