+++ tests/compiler/std-wasi/crypto.debug.wat
@@ -13,9 +13,9 @@
  (type $i32_i64_i32_=>_none (func (param i32 i64 i32)))
  (type $i64_i32_=>_i32 (func (param i64 i32) (result i32)))
  (type $i32_i64_i32_i32_=>_none (func (param i32 i64 i32 i32)))
- (import "wasi_snapshot_preview1" "fd_write" (func $~lib/bindings/wasi_snapshot_preview1/fd_write (param i32 i32 i32 i32) (result i32)))
- (import "wasi_snapshot_preview1" "proc_exit" (func $~lib/bindings/wasi_snapshot_preview1/proc_exit (param i32)))
- (import "wasi_snapshot_preview1" "random_get" (func $~lib/bindings/wasi_snapshot_preview1/random_get (param i32 i32) (result i32)))
+ (import "wasi_snapshot_preview1" "fd_write" (func $~lib/bindings/internal/wasi_snapshot_preview1/fd_write (param i32 i32 i32 i32) (result i32)))
+ (import "wasi_snapshot_preview1" "proc_exit" (func $~lib/bindings/internal/wasi_snapshot_preview1/proc_exit (param i32)))
+ (import "wasi_snapshot_preview1" "random_get" (func $~lib/bindings/internal/wasi_snapshot_preview1/random_get (param i32 i32) (result i32)))
  (global $~lib/shared/runtime/Runtime.Stub i32 (i32.const 0))
  (global $~lib/shared/runtime/Runtime.Minimal i32 (i32.const 1))
  (global $~lib/shared/runtime/Runtime.Incremental i32 (i32.const 2))
@@ -36,7 +36,7 @@
  (global $std-wasi/crypto/ab (mut i32) (i32.const 0))
  (global $std-wasi/crypto/buf (mut i32) (i32.const 0))
  (global $~lib/native/ASC_NO_ASSERT i32 (i32.const 0))
- (global $~lib/bindings/wasi_errors/ERRNO_TO_STRING i32 (i32.const 3648))
+ (global $~lib/bindings/internal/wasi_errors/ERRNO_TO_STRING i32 (i32.const 3648))
  (global $~lib/process/process.stdout i32 (i32.const 1))
  (global $~lib/bindings/wasi/tempbuf i32 (i32.const 5984))
  (global $~lib/builtins/i32.MAX_VALUE i32 (i32.const 2147483647))
@@ -159,7 +159,7 @@
  (elem $0 (i32.const 1))
  (export "memory" (memory $0))
  (export "_start" (func $~start))
- (func $~lib/bindings/wasi_snapshot_preview1/iovec#set:buf (param $0 i32) (param $1 i32)
+ (func $~lib/bindings/internal/wasi_snapshot_preview1/iovec#set:buf (param $0 i32) (param $1 i32)
   local.get $0
   local.get $1
   i32.store
@@ -654,7 +654,7 @@
   end
   unreachable
  )
- (func $~lib/bindings/wasi_snapshot_preview1/iovec#set:buf_len (param $0 i32) (param $1 i32)
+ (func $~lib/bindings/internal/wasi_snapshot_preview1/iovec#set:buf_len (param $0 i32) (param $1 i32)
   local.get $0
   local.get $1
   i32.store offset=4
@@ -666,7 +666,7 @@
   (local $7 i32)
   i32.const 0
   i32.const 12
-  call $~lib/bindings/wasi_snapshot_preview1/iovec#set:buf
+  call $~lib/bindings/internal/wasi_snapshot_preview1/iovec#set:buf
   i32.const 12
   local.set $4
   local.get $4
@@ -806,15 +806,15 @@
   local.get $4
   i32.const 12
   i32.sub
-  call $~lib/bindings/wasi_snapshot_preview1/iovec#set:buf_len
+  call $~lib/bindings/internal/wasi_snapshot_preview1/iovec#set:buf_len
   i32.const 2
   i32.const 0
   i32.const 1
   i32.const 8
-  call $~lib/bindings/wasi_snapshot_preview1/fd_write
+  call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
   drop
   i32.const 255
-  call $~lib/bindings/wasi_snapshot_preview1/proc_exit
+  call $~lib/bindings/internal/wasi_snapshot_preview1/proc_exit
  )
  (func $~lib/rt/itcms/Object#set:nextWithColor (param $0 i32) (param $1 i32)
   local.get $0
@@ -2988,7 +2988,7 @@
   i32.add
   local.get $0
   i32.load offset=8
-  call $~lib/bindings/wasi_snapshot_preview1/random_get
+  call $~lib/bindings/internal/wasi_snapshot_preview1/random_get
   local.set $1
   i32.const 0
   drop
@@ -2997,7 +2997,7 @@
   i32.and
   if
    local.get $1
-   call $~lib/bindings/wasi_errors/errnoToString
+   call $~lib/bindings/internal/wasi_errors/errnoToString
    i32.const 4032
    i32.const 17
    i32.const 18
@@ -3675,14 +3675,14 @@
     i32.const 4
     i32.mul
     i32.add
-    call $~lib/bindings/wasi_snapshot_preview1/fd_write
+    call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
     local.set $7
     local.get $7
     i32.const 65535
     i32.and
     if
      local.get $7
-     call $~lib/bindings/wasi_errors/errnoToString
+     call $~lib/bindings/internal/wasi_errors/errnoToString
      i32.const 6032
      i32.const 211
      i32.const 16
@@ -3732,7 +3732,7 @@
   i32.const 4
   i32.mul
   i32.add
-  call $~lib/bindings/wasi_snapshot_preview1/fd_write
+  call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
   local.set $10
   local.get $9
   call $~lib/rt/tlsf/__free
@@ -3741,7 +3741,7 @@
   i32.and
   if
    local.get $10
-   call $~lib/bindings/wasi_errors/errnoToString
+   call $~lib/bindings/internal/wasi_errors/errnoToString
    i32.const 6032
    i32.const 222
    i32.const 12
@@ -3865,7 +3865,7 @@
   i32.const 5872
   local.get $0
   call $~lib/rt/itcms/__visit
-  global.get $~lib/bindings/wasi_errors/ERRNO_TO_STRING
+  global.get $~lib/bindings/internal/wasi_errors/ERRNO_TO_STRING
   local.tee $1
   if
    local.get $1
@@ -3990,7 +3990,7 @@
    unreachable
   end
  )
- (func $~lib/bindings/wasi_errors/errnoToString (param $0 i32) (result i32)
+ (func $~lib/bindings/internal/wasi_errors/errnoToString (param $0 i32) (result i32)
   (local $1 i32)
   global.get $~lib/memory/__stack_pointer
   i32.const 4
@@ -4006,7 +4006,7 @@
   i32.const 77
   i32.lt_u
   if
-   global.get $~lib/bindings/wasi_errors/ERRNO_TO_STRING
+   global.get $~lib/bindings/internal/wasi_errors/ERRNO_TO_STRING
    local.set $1
    global.get $~lib/memory/__stack_pointer
    local.get $1
