+++ tests/compiler/wasi/seed.debug.wat
@@ -5,7 +5,7 @@
  (type $i32_=>_i32 (func (param i32) (result i32)))
  (type $i64_=>_none (func (param i64)))
  (type $none_=>_none (func))
- (import "wasi_snapshot_preview1" "random_get" (func $~lib/bindings/wasi_snapshot_preview1/random_get (param i32 i32) (result i32)))
+ (import "wasi_snapshot_preview1" "random_get" (func $~lib/bindings/internal/wasi_snapshot_preview1/random_get (param i32 i32) (result i32)))
  (global $~lib/math/random_seeded (mut i32) (i32.const 0))
  (global $~lib/bindings/wasi/tempbuf i32 (i32.const 16))
  (global $~lib/math/random_state0_64 (mut i64) (i64.const 0))
@@ -27,7 +27,7 @@
   loop $do-loop|0
    global.get $~lib/bindings/wasi/tempbuf
    i32.const 8
-   call $~lib/bindings/wasi_snapshot_preview1/random_get
+   call $~lib/bindings/internal/wasi_snapshot_preview1/random_get
    drop
    global.get $~lib/bindings/wasi/tempbuf
    i64.load
