+++ tests/compiler/wasi/trace.release.wat
@@ -10,8 +10,8 @@
  (type $i32_i32_=>_i32 (func (param i32 i32) (result i32)))
  (type $i32_i64_i64_i32_i64_i32_=>_i32 (func (param i32 i64 i64 i32 i64 i32) (result i32)))
  (type $i32_f64_f64_f64_f64_f64_=>_none (func (param i32 f64 f64 f64 f64 f64)))
- (import "wasi_snapshot_preview1" "fd_write" (func $~lib/bindings/wasi_snapshot_preview1/fd_write (param i32 i32 i32 i32) (result i32)))
- (import "wasi_snapshot_preview1" "proc_exit" (func $~lib/bindings/wasi_snapshot_preview1/proc_exit (param i32)))
+ (import "wasi_snapshot_preview1" "fd_write" (func $~lib/bindings/internal/wasi_snapshot_preview1/fd_write (param i32 i32 i32 i32) (result i32)))
+ (import "wasi_snapshot_preview1" "proc_exit" (func $~lib/bindings/internal/wasi_snapshot_preview1/proc_exit (param i32)))
  (global $~lib/rt/tlsf/ROOT (mut i32) (i32.const 0))
  (global $~argumentsLength (mut i32) (i32.const 0))
  (global $~lib/util/number/_frc_plus (mut i64) (i64.const 0))
@@ -439,10 +439,10 @@
   i32.const 0
   i32.const 1
   i32.const 8
-  call $~lib/bindings/wasi_snapshot_preview1/fd_write
+  call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
   drop
   i32.const 255
-  call $~lib/bindings/wasi_snapshot_preview1/proc_exit
+  call $~lib/bindings/internal/wasi_snapshot_preview1/proc_exit
  )
  (func $~lib/rt/tlsf/removeBlock (param $0 i32) (param $1 i32)
   (local $2 i32)
@@ -2676,7 +2676,7 @@
   local.get $6
   i32.const 1
   local.get $7
-  call $~lib/bindings/wasi_snapshot_preview1/fd_write
+  call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
   drop
   i32.const 1052
   i32.load
@@ -2695,7 +2695,7 @@
   local.get $6
   i32.const 1
   local.get $7
-  call $~lib/bindings/wasi_snapshot_preview1/fd_write
+  call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
   drop
   local.get $0
   if (result i32)
@@ -2723,7 +2723,7 @@
    local.get $6
    i32.const 1
    local.get $7
-   call $~lib/bindings/wasi_snapshot_preview1/fd_write
+   call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
    drop
    local.get $0
    i32.const 1
@@ -2747,7 +2747,7 @@
     local.get $6
     i32.const 1
     local.get $7
-    call $~lib/bindings/wasi_snapshot_preview1/fd_write
+    call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
     drop
     local.get $0
     i32.const 2
@@ -2771,7 +2771,7 @@
      local.get $6
      i32.const 1
      local.get $7
-     call $~lib/bindings/wasi_snapshot_preview1/fd_write
+     call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
      drop
      local.get $0
      i32.const 3
@@ -2795,7 +2795,7 @@
       local.get $6
       i32.const 1
       local.get $7
-      call $~lib/bindings/wasi_snapshot_preview1/fd_write
+      call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
       drop
       local.get $0
       i32.const 4
@@ -2819,7 +2819,7 @@
        local.get $6
        i32.const 1
        local.get $7
-       call $~lib/bindings/wasi_snapshot_preview1/fd_write
+       call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
        drop
       end
      end
@@ -2840,7 +2840,7 @@
   local.get $6
   i32.const 1
   local.get $7
-  call $~lib/bindings/wasi_snapshot_preview1/fd_write
+  call $~lib/bindings/internal/wasi_snapshot_preview1/fd_write
   drop
   local.get $6
   i32.const 19024
